Move FilterStatus to useFilterRounds
